+++ packages/isaacscript-common/src/types/WeightedArray.ts
@@ -1,2 +1,2 @@
-/** An array where each element is paired with a number indicating that elements 'weight'. */
+/** An array where each element is paired with a number indicating that element's weight. */
 export type WeightedArray<T> = Array<[T, float]>;