+++ packages/docs/docs/main/change-log.md
@@ -15,6 +15,10 @@ This page lists the changes to the IsaacScript framework.
   - `isGridEntityXMLType` (for the `PRE_ROOM_ENTITY_SPAWN` callback)
 - Renamed the following helper functions:
   - `isPoopGridEntityType` --> `isPoopGridEntityXMLType`
+- Added the following helper types:
+  - `ERange`
+- Renamed the following helper types:
+  - `Range` --> `IRange`
 
 ## September 18th, 2023